增加 部分测试代码

+++ biz-base/src/main/java/com/biz/iems/mall/eo/StoreAddrEo.java
@@ -32,5 +32,5 @@ public class StoreAddrEo implements Serializable {
 
     private String codePrefix;
 
-    private Integer codeSerialNum;
+    private Long codeSerialNum;
 }
